requirements.txt
cython
numpy
pyproj
cartopy==0.17.0
matplotlib==3.0.1
geopandas==0.4.0
geopy==1.17.0
requests
pendulum
boto3
pytest
